+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added timeout to runner call
 * Added loaded file name to window title
 * Added optional input scaling for coordinate list inputs & compressed svg output
 * Added autozoom functionality (not yet fully complete/tested)